requirements.txt
langchain
langchain-google-genai
python-dotenv
requests
beautifulsoup4
lxml
pandas
playwright
google-genai
